refactor: separate `FallbackTranslationRepository` interface and implementation
* Convert `FallbackTranslationRepository` into an interface and move its logic to `DefaultFallbackTranslationRepository`.
* Update `FallbackTranslationRepository.getTranslation` signature to accept `providerName` and `apiKey`.
* Update DI configuration to bind `FallbackTranslationRepository` to its new implementation.
